add documentation, don't worry! more where relevant is to come!
Here is the same thing from ChatGPT:

Hark! Attend thee, fair reader, and hearken unto my words. Forsooth, I
bid thee to take note and heed my counsel, for though my pen be but a
humble instrument, yet it doth possess the power to inspire and
enlighten.
Therefore, let it be known that henceforth, documentation has been
added, without fear or hesitation, and be not dismayed, for more
where relevant is to come!